+++ Thousand.Core/Model/Font.cs
@@ -1,6 +1,6 @@
 ﻿namespace Thousand.Model
 {
-    public record Font(string Family, int Size, Colour Colour)
+    public record Font(string Family, decimal Size, Colour Colour)
     {
         public Font() : this(SkiaSharp.SKTypeface.Default.FamilyName, 20, Colour.Black) { }
     }
